Added Changes: Simplied Installation. Slight change to keyboard layout. Reformed README.md

+++ remover.sh
@@ -1,7 +1,17 @@
 #!/bin/sh
 
-sudo rmmod bbqX0kbd
-sudo rm -rf /boot/overlays/i2c-bbqX0kbd.dtbo
-sudo rm -rf /lib/modules/$(uname -r)/kernel/drivers/i2c/bbqX0kbd.ko
-make clean
+if sudo rmmod bbqX0kbd ; then
 
+	sudo sed -i '/bbqX0kbd/d' /boot/config.txt
+	sudo rm -rf /boot/overlays/i2c-bbqX0kbd.dtbo
+
+	sudo sed -i '/bbqX0kbd/d' /etc/modules
+	sudo rm -rf /lib/modules/$(uname -r)/kernel/drivers/i2c/bbqX0kbd.ko
+
+	sudo sed -i '/bbq10kbd/d' /etc/default/keyboard
+	sudo rm -rf /usr/local/share/kbd/keymaps/bbq10kbd.map
+
+	make clean
+else
+	echo "BBQX0KBD Driver is not being used.\nIf it was removed manually, remove contents from  files and locations manually.\nRefer to the contents of remover.sh to find what to remove from where.\n\n"
+fi
